Fix/added-death-screen
I have added the death screen when the health depletes. I put in a 2 second delay so we can see the animation of the dog laying down.
I am still getting the error
 'UnassignedReferenceException: The variable gameOverScreen of Health has not been assigned.
You probably need to assign the gameOverScreen variable of the Health script in the inspector.
Health.Awake () (at Assets/Scripts/Health.cs:28)'